fix: moves files to follow the import rules
