+++ src/hooks/__tests__/useReleaseStatusOnFavicon.spec.ts
@@ -1,15 +1,15 @@
 import { global_success_color_100 as greenColor } from '@patternfly/react-tokens/dist/js/global_success_color_100';
 import { renderHook } from '@testing-library/react';
 import { runStatus } from '~/consts/pipelinerun';
-import * as statusColorUtils from '~/shared/utils/status-color-utils';
-import * as useFaviconStatusBadgeModule from '../useFaviconStatusBadge';
+import * as useFaviconStatusBadgeModule from '~/shared/hooks/useFaviconStatusBadge';
+import * as statusColorUtils from '~/utils/status-color-utils';
 import { useReleaseStatusOnFavicon } from '../useReleaseStatusOnFavicon';
 
-jest.mock('../useFaviconStatusBadge', () => ({
+jest.mock('~/shared/hooks/useFaviconStatusBadge', () => ({
   useFaviconStatusBadge: jest.fn(),
 }));
 
-jest.mock('~/shared/utils/status-color-utils', () => ({
+jest.mock('~/utils/status-color-utils', () => ({
   getColorForPipelineStatus: jest.fn(),
   getColorForReleaseStatus: jest.fn(),
 }));
