+++ src/hooks/usePipelineRunStatusOnFavicon.ts
@@ -1,6 +1,6 @@
 import { runStatus } from '~/consts/pipelinerun';
-import { getColorForPipelineStatus } from '~/shared/utils/status-color-utils';
-import { useFaviconStatusBadge } from './useFaviconStatusBadge';
+import { useFaviconStatusBadge } from '~/shared/hooks/useFaviconStatusBadge';
+import { getColorForPipelineStatus } from '~/utils/status-color-utils';
 
 /** Updates the favicon badge based on a pipeline run status. */
 export const usePipelineRunStatusOnFavicon = (status: runStatus): void => {
