+++ src/shared/hooks/index.ts
@@ -7,5 +7,3 @@ export * from './useDeepCompareMemoize';
 export * from './useMutationObserver';
 export * from './useLazyActionMenu';
 export * from './useFaviconStatusBadge';
-export * from './usePipelineRunStatusOnFavicon';
-export * from './useReleaseStatusOnFavicon';
